[deps] require 'regex' module
The regex module extends the builtin re module in a number of useful ways,
and is actually meant to be included with python at some point.

Firstly, it makes it possible to have multiple named capture groups with
the same name, greatly simplifying group parsing. It also has support for
releasing the GIL during matching and some other nifty performance improvements.

Currently, it is only used in the reminders module, which required it because
of the named capture group feature.

Support for using regexes compiled with the regex module in command regexes
was added in 2d201bdfcbb7452ef34df28761b300f54f094985 - instead of passing
a string, just pass the compiled regex object.