Add unit testing and static analysis support
This adds support for unit testing through the `unittest` framework
and static analysis through use of pyflakes. Another addition is the
pre-commit.sh bash script. This contains the infrastructure for running
tests and static analysis. This is intended to be run before anything
is committed to the repo - please use it by symlinking it to .git/hooks
as follows:

    ln -s ../../pre-commit.sh .git/hooks/pre-commit